+++ src/app/components/exper-lab/exper-lab.component.ts
@@ -37,7 +37,9 @@ export class ExperLabComponent implements OnInit{
 
   public delete(id?: number, url?: string): void {
     if(id != undefined) {
-      this.uploadService.deleteFileByUrl(url);
+      if(url.startsWith('https://firebasestorage')){
+        this.uploadService.deleteFileByUrl(url);
+      }
       this.experLabService.delete(id).subscribe({
         next: (res) => {
           console.log("Experiencia eliminada");
